requirements.txt
fastapi
uvicorn[standard]
httpx
meilisearch
beautifulsoup4
SQLAlchemy
apscheduler
langchain
langchain-community
langchain-openai
langchain-core
chromadb
pydantic
python-dotenv
torch
transformers
jieba
scikit-learn
fpdf2
